Finish README
Let's recover some of the information, I've already learnt. 
Git is a web-service for public hosting and developing of IT-projects, so the people can share, edit and discuss their projects. 
to use it you need to follow some steps 
1. create a repository 
that's a database that collects all the information that you need to know before opening and using, modifying the project 
2. create a branch 
it allows you and others to modify projects and then adding them to a master branch 
3. then make edits in your branch 
then it contains content that's different from the master branch